tests: Move and rename test shader templates
